requirements.txt
Django==4.2.7
gunicorn==21.2.0
djangorestframework==3.14.0
jdatetime==4.1.1
psycopg2==2.9.9
